change ids from integer to long

+++ backend/src/main/java/com/library/catalog/models/Book.java
@@ -2,13 +2,11 @@ package com.library.catalog.models;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.Hibernate;
 
 import java.util.List;
 
@@ -24,7 +22,7 @@ public class Book {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true)
-    private Integer id;
+    private Long id;
 
     @Column(name = "title", unique = true, nullable = false)
     @NotBlank
